\name{MKinfer-package}
\alias{MKinfer-package}
\alias{MKinfer}
\docType{package}
\title{
Inferential Statistics.
}
\description{
Computation of various confidence intervals (Altman et al. (2000), ISBN:978-0-727-91375-3; 
Hedderich and Sachs (2018), ISBN:978-3-662-56657-2) including bootstrapped 
versions (Davison and Hinkley (1997), ISBN:978-0-511-80284-3) as well as 
Hsu (Hedderich and Sachs (2018), ISBN:978-3-662-56657-2), 
permutation (Janssen (1997), <doi:10.1016/S0167-7152(97)00043-6>), 
bootstrap (Davison and Hinkley (1997), ISBN:978-0-511-80284-3) and 
multiple imputation (Barnard and Rubin (1999), <doi:10.1093/biomet/86.4.948>) t-test.
}
\details{
\tabular{ll}{
Package: \tab MKinfer\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2019-11-27\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats, MKdescr, boot, arrangements\cr
Suggests: \tab Amelia, knitr, rmarkdown, exactRankTests\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKinfer)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
