\name{MKinfer-package}
\alias{MKinfer-package}
\alias{MKinfer}
\docType{package}
\title{
Inferential Statistics.
}
\description{
Computation of various confidence intervals (Altman et al. (2000), ISBN:978-0-727-91375-3; 
Hedderich and Sachs (2018), ISBN:978-3-662-56657-2) including bootstrapped 
versions (Davison and Hinkley (1997), ISBN:978-0-511-80284-3) as well as 
Hsu (Hedderich and Sachs (2018), ISBN:978-3-662-56657-2), 
permutation (Janssen (1997), <doi:10.1016/S0167-7152(97)00043-6>), 
bootstrap (Davison and Hinkley (1997), ISBN:978-0-511-80284-3) and 
multiple imputation (Barnard and Rubin (1999), <doi:10.1093/biomet/86.4.948>) t-test
and Wilcoxon tests. Graphical visualization by volcano and Bland-Altman plots 
(Bland and Altman (1986), <doi:10.1016/S0140-6736(86)90837-8>; Shieh (2018), <doi:10.1186/s12874-018-0505-y>).
}
\details{
library(MKinfer)
}
\author{ Matthias Kohl \url{https://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
