\name{MKinfer-package}
\alias{MKinfer-package}
\alias{MKinfer}
\docType{package}
\title{
Inferential Statistics.
}
\description{
Functions not available in standard R packages such as computation of various confidence intervals including bootstrapped versions as well as HSU, permutation, bootstrap and multiple imputation t-test.
}
\details{
\tabular{ll}{
Package: \tab MKinfer\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2019-11-22\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats, MKdescr, boot, arrangements\cr
Suggests: \tab Amelia, knitr, rmarkdown, exactRankTests\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKinfer)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
